implement very basic projectile motion animation
Currently only works for when initial vertical displacement is 0. The
way it works is that it maintains its own local coordinates for the
ball starting at (0, 0). It uses motion equations to calculate its next
coordinates and updates the local coordinates. In order to move the
ball, the function takes a look at the differences between the previous
coordinates and the new coordinates, telling it how much it should move
in each axis. This difference is multiplied by some number that
determines how much a given unit (m, ft, eg.., no units rn though)
should be expressed in pixels. Thanks to the canvas.move() function,
I don't need to worry about the ball's actual coordinates in the canvas
(for now..., but forever I hope)