feat: add button to reset ball position
Added a button that will bring ball back to original position after
it has moved after a projectile motion animation.

Also fixed a bug that probably happens due to the nature of floating
point arithmetic. Sometimes (eg. initial velocity 20, angle 76), the
time steps would not add up exactly back to the original total time,
going negligibly higher. This would cause the last time not to be
computed and an error in the animation, where the ball would stop much
higher than it should be. Currently the fix is hardcoded for when time
steps are a hundredth of the total time, but that will change.