added a distance unit combobox
Allows user to pick what unit they want to use in their animation. No
use as of this commit but may be useful later.

Used the more appropriate Spinbox widget to collect initial velocity and
angle options. Made coordinates for the top of the floor (floor_h) and
the unit to pixel conversion number (upc_num) instance level variables
as they will be useful for future plans.